refactors fixture
